+++ test-app/src/main/java/android/template/test/app/testdi/TestDatabaseModule.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package android.template.test.navigation.testdi
+package android.template.test.app.testdi
 
 import dagger.Binds
 import dagger.Module
